+++ src/test/java/step03/domain/NameTest.java
@@ -1,4 +1,4 @@
-package step03;
+package step03.domain;
 
 import exception.OutOfNameLengthException;
 import org.junit.jupiter.api.DisplayName;
