+++ src/test/java/step03/domain/ParticipantsTest.java
@@ -1,4 +1,4 @@
-package step03;
+package step03.domain;
 
 import exception.DuplicatedParticipantsNameException;
 import exception.InvalidCountOfParticipantsException;
